+++ rollup.config.js
@@ -1,13 +1,9 @@
-import autoprefixer from 'autoprefixer'
-import sass from 'node-sass'
-
 import babel from 'rollup-plugin-babel'
-import clean from 'postcss-clean'
 import commonjs from 'rollup-plugin-commonjs'
 import copy from 'rollup-plugin-copied'
 import external from 'rollup-plugin-peer-deps-external'
-import postcss from 'rollup-plugin-postcss'
 import resolve from 'rollup-plugin-node-resolve'
+import scss from 'rollup-plugin-scss'
 import url from 'rollup-plugin-url'
 
 import pkg from './package.json'
@@ -26,10 +22,8 @@ export default {
   ],
   plugins: [
     external(),
-    postcss({
-      plugins: [autoprefixer, clean],
-      sourceMap: true,
-      extract: 'dist/css/bundle.min.css'
+    scss({
+      output: './dist/css/bundle.min.css'
     }),
     url(),
     babel({